refactor: studentLayout에서 container 분리
